rename files and update doc
